Rename nipio.js to service-host.js
