+++ packages/navy/src/cli/ps.js
@@ -3,7 +3,7 @@
 import chalk from 'chalk'
 
 import {getNavy} from '../'
-import {getUrlFromService} from '../util/nipio'
+import {getUrlFromService} from '../util/service-host'
 import table from '../util/table'
 
 const SMALL_WINDOW_COLUMNS = 185
