+++ packages/navy/src/middleware/add-virtual-hosts.js
@@ -2,7 +2,7 @@
 
 import {find} from 'lodash'
 import {Navy} from '../navy'
-import {createHostForService} from '../util/nipio'
+import {createHostForService} from '../util/service-host'
 
 const getServiceHTTPProxyConfig = (serviceName, navyFile) => {
   if (navyFile && navyFile.httpProxy && navyFile.httpProxy[serviceName]) {
